+++ ConnectorNavisworks/ConnectorNavisworks/Bindings/ConnectorNavisworksBindings.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using Autodesk.Navisworks.Api;
 using DesktopUI2;
 using DesktopUI2.Models.Settings;
 using Speckle.Core.Kits;
@@ -11,30 +12,18 @@ namespace Speckle.ConnectorNavisworks.Bindings
 {
     public partial class ConnectorBindingsNavisworks : ConnectorBindings
     {
-      
+        // Much of the interaction in Navisworks is through the ActiveDocument API
+        public static Document Doc;
+
+        // Majority of interaction with Speckle will be through the saved selection and search Sets
+        public SavedItemCollection SavedSets { get; set; }
+
         public ConnectorBindingsNavisworks(Document navisworksActiveDocument) : base()
         {
             Doc = navisworksActiveDocument;
+            SavedSets = Doc.SelectionSets.ToSavedItemCollection();
         }
 
-
-        //private static readonly string ApplicationIdKey = "applicationId";
-        public static Document Doc;
-
-
-        public override List<ReceiveMode> GetReceiveModes()
-        {
-            // No receive modes available
-            return new List<ReceiveMode>
-            {
-                ReceiveMode.Create
-            };
-        }
-
-        /// <summary>
-        /// Stored Base objects from commit flattening on receive: key is the Base id
-        /// </summary>
-        public Dictionary<string, Base> StoredObjects = new Dictionary<string, Base>();
 
         public override string GetActiveViewName() => "Entire Document";
 
@@ -45,14 +34,8 @@ namespace Speckle.ConnectorNavisworks.Bindings
         }
 
 
-        public override string GetHostAppName() => Utils.Slug;
+        public override string GetHostAppName() => HostApplications.Navisworks.Slug;
 
-
-        public override List<string> GetObjectsInView() // this returns all visible doc objects.
-        {
-            List<string> objects = new List<string>();
-            return objects;
-        }
 
         public override string GetHostAppNameVersion() =>
             Utils.VersionedAppName.Replace("Navisworks", "Navisworks "); //hack for ADSK store;
@@ -66,7 +49,7 @@ namespace Speckle.ConnectorNavisworks.Bindings
 
         public override List<ISetting> GetSettings() => new List<ISetting> { };
 
-        private string GetDocPath(Document doc) => "";
+        private static string GetDocPath(Document doc) => "";
 
         public override string GetDocumentLocation() => GetDocPath(Doc);
